Restore checkpoint: Use cli to drive execution
Instead of keeping track of the available checkpoints in the farcaster
runtime, we use the cli to drive execution by first retrieving the
checkpoint information for a certain swap id and then triggering the
restore workflow.

Because the checkpoint list does not need to be persisted in farcasterd
anymore, the checkpoint list is no longer routed through farcasterd.

Update the docs to reflect the change.